Update Index.html
Added a favicon link in the <head> section.
Added aria-label to the <nav> element to provide a more accessible experience for users with screen readers.
Added aria-current="page" to the current active link in the navigation menu. This helps with accessibility. 
 Added a role="button" attribute to the "Shop Now" link in the header. This attribute helps indicate to screen readers and other assistive technologies that this link acts as a button.
Added aria-labelledby attributes to the <section> elements for "Featured Artworks" and "About Us" to improve accessibility. This attribute associates the sections with their respective headings.
Added id attributes to the headings of the "Featured Artworks" and "About Us" sections. These id attributes are used in combination with the aria-labelledby attributes to establish relationships between the sections and their headings, providing a better experience for users with screen readers.