Create EtsyInsightV2.py
We added a new function called download_listing_data that takes a Listing object and the name of a store as inputs, and downloads the listing's data (including the image) and saves it to a folder named after the store and the product title.
We updated the get_shop_listings function to call the new download_listing_data function for each listing, passing in the listing and the store name.
We also added a new argument to the get_shop_listings function called max_listings_per_shop. This argument allows us to limit the number of listings we download for each store to a specified number. We set the default value to 5, so that the program only downloads the top 5 listings for each store.
Finally, we added some code to create the necessary folders if they don't already exist.
These updates allow us to download the images and descriptions for each of the top 5 listings for each store and save them to folders named after the store and product title. This can be useful for further analysis or for uploading the products to our own store with our own product descriptions.